Fix for contact us and Navigation
