+++ src/app/about/about.component.scss
@@ -1,4 +1,3 @@
-@import "../../assets/scss/landing.scss";
 section {
     padding-top: 0;
     margin-bottom: 50px;
