+++ src/app/contact/contact.component.scss
@@ -1,4 +1,3 @@
-@import "../../assets/scss/landing.scss";
 section {
   padding-top: 0;
   margin-bottom: 50px;
@@ -37,26 +36,3 @@ section {
   }
 }
 
-.has-error input,
-.has-error select,
-.has-error textarea {
-  border: 1px solid #b32b23;
-}
-
-.error-message {
-  color: #b32b23;
-  font-size: 14px;
-  margin-top: 5px;
-}
-
-.email-thankyou {
-    height: 500px;
-}
-
-.message-textarea {
-    margin-top: 10px;
-}
-.text-info {
-  color: #1160cc;
-  text-align: center;
-}
