+++ src/app/landing/client-section/client-section.component.html
@@ -1,5 +1,5 @@
 <!--brand slider start-->
-<section class="app2 brand-sliders">
+<section class="app2 brand-sliders bg-light-inner">
     <div class="container">
         <div class="row">
             <div class="col-md-10 offset-md-1">
