+++ src/app/landing/contact-section/contact-section.component.html
@@ -1,5 +1,6 @@
 <!-- Contact form start -->
-<section class=" bg-light-inner" id="contact">
+
+<section id="contact">
     <div class="container">
         <div class="row">
             <div class="col-12">
@@ -25,66 +26,7 @@
     </div>
     <section class="music bg-footer">
         <div class="container">
-            <div class="form form-footer p-t-50" *ngIf="!submitted">
-                <form [formGroup]="contactForm" (ngSubmit)="onSubmit()">
-                    <div class="row">
-                        <div class="col-sm-6">
-                            <div class="form-group" [ngClass]="{ 'has-error': contactForm.get('name')?.invalid && (contactForm.get('name')?.dirty || contactForm.get('name')?.touched) }">
-                                <input formControlName="name" type="text" placeholder="Name" id="name">
-                                <div class="error-message" *ngIf="contactForm.get('name')?.invalid && (contactForm.get('name')?.dirty || contactForm.get('name')?.touched)">
-                                    Field is required.
-                                  </div>
-                            </div>
-                        </div>
-                        <div class="col-sm-6">
-                            <div class="form-group" [ngClass]="{ 'has-error': contactForm.get('email')?.invalid && (contactForm.get('email')?.dirty || contactForm.get('email')?.touched) }">
-                                <input formControlName="email" type="text" placeholder="Email" id="Email">
-                                <div class="error-message" *ngIf="contactForm.get('email')?.invalid && (contactForm.get('email')?.dirty || contactForm.get('email')?.touched)">
-                                    <span *ngIf="contactForm.get('email').hasError('required')">Field is required.</span>
-                                    <span *ngIf="contactForm.get('email').hasError('email')">Enter a valid email ID.</span>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="col-sm-6">
-                            <div class="form-group">
-                                <select name="country" id="country">
-                                    <option value="" disabled selected>Select State</option>
-                                    <option *ngFor="let country of countries" [value]="country.name">{{country.name}}
-                                    </option>
-                                </select>
-                            </div>
-                        </div>
-                        <div class="col-sm-6">
-                            <div class="form-group" [ngClass]="{ 'has-error': contactForm.get('services')?.invalid && (contactForm.get('services')?.dirty || contactForm.get('services')?.touched) }">
-                                <select name="service" id="service" formControlName="services">
-                                    <option value="" disabled selected>Select service</option>
-                                    <option *ngFor="let service of servicesProvided" [value]="service.name">{{service.name}}
-                                    </option>
-                                </select>
-                                <div class="error-message" *ngIf="contactForm.get('services')?.invalid && (contactForm.get('services')?.dirty || contactForm.get('services')?.touched)">
-                                    Field is required.
-                                  </div>
-                            </div>
-                        </div>
-                        <div class="col-12">
-                            <div class="form-group" [ngClass]="{ 'has-error': contactForm.get('message')?.invalid && (contactForm.get('message')?.dirty || contactForm.get('message')?.touched) }">
-                                <textarea class="form-control" formControlName="message" rows="6"
-                                    placeholder="Please provide details about your project, including description and timeline." pInputTextarea [autoResize]="true"></textarea>
-                                    <div class="error-message"
-                                    *ngIf="contactForm.get('message')?.invalid && (contactForm.get('message')?.dirty || contactForm.get('message')?.touched)">
-                                    Field is required.
-                                </div>
-                            </div>
-                        </div>
-                        <button type="submit" class="btn btn-default btn-gradient text-white">Send</button>
-                    </div>
-                </form>
-            </div>
-            <div class="col-xl-5 p-l-0 col-md-6 set-z-index form-footer" *ngIf="submitted">
-                <div class="bg-white email-thankyou">
-                    <h3 class="text-info">Thank you for submitting your contact information! We will get back to you as soon as possible.</h3>
-                </div>
-            </div>
+            <app-contact-us></app-contact-us>
         </div>
     </section>
     <img src="assets/images/event/footer/2.png" alt="" class="set-abs top-0 plane2">
