--- src/app/landing/landing.component.scss
@@ -1,1 +1,0 @@
-@import "../../assets/scss/landing.scss";
