+++ src/app/portfolio/portfolio.component.scss
@@ -1,5 +1,3 @@
-@import "../../assets/scss/landing.scss";
-
 .container {
     max-width: 1140px;
   }
