+++ src/app/shared/components/contact-us/contact-us.component.scss
@@ -1,0 +1,23 @@
+.has-error input,
+.has-error select,
+.has-error textarea {
+  border: 1px solid #b32b23;
+}
+
+.error-message {
+  color: #b32b23;
+  font-size: 14px;
+  margin-top: 5px;
+}
+
+.email-thankyou {
+    height: 500px;
+}
+
+.message-textarea {
+    margin-top: 10px;
+}
+.text-info {
+  color: #1160cc;
+  text-align: center;
+}
