+++ src/app/shared/components/header/header.component.html
@@ -17,19 +17,19 @@
                             <div class="collapse navbar-collapse" id="scroll-spy">
                                 <ul class="navbar-nav mx-auto nav">
                                     <li class="nav-item">
-                                        <a class="nav-link" [routerLink]="['/landing']">Home</a>
+                                      <a class="nav-link" [routerLink]="['/landing']" routerLinkActive="active">Home</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a class="nav-link"  [routerLink]="['/aboutus']">About Us</a>
+                                        <a class="nav-link"  [routerLink]="['/aboutus']" routerLinkActive="active">About Us</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a class="nav-link"  [routerLink]="['/service']">Services</a>
+                                        <a class="nav-link"  [routerLink]="['/service']" routerLinkActive="active">Services</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a class="nav-link" [routerLink]="['/portfolio']" [ngClass]="{'active show': activeClass == 'portfolio' }">Portfolio</a>
+                                        <a class="nav-link" [routerLink]="['/portfolio']" routerLinkActive="active">Portfolio</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a class="nav-link" [routerLink]="['/contact']">Contact</a>
+                                        <a class="nav-link" [routerLink]="['/contact']" routerLinkActive="active">Contact</a>
                                     </li>
                                 </ul>
                             </div>
