+++ src/app/shared/components/header/header.component.ts
@@ -7,26 +7,9 @@ import { NavigationEnd, Router } from '@angular/router';
   styleUrls: ['./header.component.scss']
 })
 export class HeaderComponent implements OnInit {
-  public url : any;
-  public activeClass = 'home';
-  constructor(private router: Router) {
-    this.router.events.subscribe((event) => {
-      if (event instanceof NavigationEnd) {
-        this.url = event.url;
-      }
-    });
+  constructor() {
   }
 
   ngOnInit() {
   }
-
-  active(fragement: string) {
-    this.activeClass = fragement;
-    this.router.navigate(
-      ['/landing'],
-      { fragment: fragement }
-    );
-  }
-
-
 }
