--- src/assets/css/flaticon.css
@@ -1,42 +1,0 @@
-/*
-    Flaticon icon font: Flaticon
-    Creation date: 07/01/2019 07:52
-    */
-@font-face {
-  font-family: "Flaticon";
-  src: url("/assets/fonts/Flaticon.eot");
-  src: url("/assets/fonts/Flaticon.eot?#iefix") format("embedded-opentype"), url("/assets/fonts/Flaticon.woff") format("woff"), url("/assets/fonts/Flaticon.ttf") format("truetype"), url("/assets/fonts/Flaticon.svg#Flaticon") format("svg");
-  font-weight: normal;
-  font-style: normal; }
-
-@media screen and (-webkit-min-device-pixel-ratio: 0) {
-  @font-face {
-    font-family: "Flaticon";
-    src: url("/assets/fonts/Flaticon.svg#Flaticon") format("svg"); } }
-
-[class^="flaticon-"]:before, [class*=" flaticon-"]:before, [class^="flaticon-"]:after, [class*=" flaticon-"]:after {
-  font-family: Flaticon;
-  font-size: 14px; }
-
-.fi:before {
-  display: inline-block;
-  font-family: "Flaticon";
-  font-style: normal;
-  font-weight: normal;
-  font-variant: normal;
-  line-height: 1;
-  text-decoration: inherit;
-  text-rendering: optimizeLegibility;
-  text-transform: none;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  font-smoothing: antialiased; }
-
-.flaticon-rocket-ship:before {
-  content: "\f100"; }
-
-.flaticon-diamond:before {
-  content: "\f101"; }
-
-.flaticon-pie-chart:before {
-  content: "\f102"; }
