+++ src/assets/scss/app.scss
@@ -14,7 +14,7 @@
 @import "themify.scss";
 @import "flaticon.scss";
 @import 'inner-page.scss';
-@import "../css/keyframes.css";
+@import "keyframes.scss";
 @import "~animate.css/animate.min.css";
 @import '/node_modules/ngx-toastr/toastr';
 @import './theme/responsive';
